% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc.R
\name{calc_proportion_ofFibers}
\alias{calc_proportion_ofFibers}
\title{Calculate the proportion of fibers in a microplastic particle collective}
\usage{
calc_proportion_ofFibers(shape, criterion = "elon")
}
\arguments{
\item{shape}{A character string to select the vector that contains either aspect ratio or elongation values.}

\item{criterion}{A character string indicating whether aspect ratio or elongation should be used to calculate the proportion. Must be either "asp" or "elon". Default is "elon".}

\item{data}{A data frame that will be manipulated.}
}
\description{
This functions computes the proportion of fibers in a microplastic particle collective.We use one of two different
shape descriptors as a criterion. For an aspect ratio larger than 3 MP is considered a fiber based on Cole (2016).
This is equal to an elongation larger than 2/3.Note that elongation must be calculated according to Crompton (2005).
}
\references{
Crompton, C., 2005. Particle Shape An Important Parameter in Pharmaceutical Manufacturing. Pharm. Manuf. Pack. Sourcer.
Cole, M., 2016. A novel method for preparing microplastic fibers. Sci. Rep. 6, 34519.
}
\author{
Uwe Schnepf
}
