% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc.R
\name{calc_width}
\alias{calc_width}
\title{Categorise particle size distributions based on their width}
\usage{
calc_width(diam)
}
\arguments{
\item{diam}{A vector containing a measure of particle size.}
}
\description{
Categorise particle size distributions based on their width
}
\references{
Merkus, H.G., 2009. Particle Size Measurements Fundamentals, Practice, Quality, 1st ed, Particle Technology Series. Springer Netherlands, Dordrecht.
}
\author{
Uwe Schnepf
}
