% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{clean_exclusion_shape}
\alias{clean_exclusion_shape}
\title{Function to excluded particles based on their shape}
\usage{
clean_exclusion_shape(data, cisp, elon = "Elongation", model)
}
\arguments{
\item{data}{A data frame that will be manipulated.}

\item{cisp}{A character string to select the vector that contains circularity or sphericity values.}

\item{elon}{A character string to select the vector that contains elongation values. Default is "Elongation".}

\item{model}{Must be either "ce" for the Circularity-Elongation model or "se" for the Sphericity-Elongation model.}
}
\description{
Implements an algorithm proposed by Kroener and Carbo to detect and delete falsely calculated shape factors
caused by digitization issues. It is based on a theoretical relationship between circularity/sphericity and elongtaion.
This relationship acts like a threshold for meausred pair of values. If they exceed the threshold, then something went
wrong in image processing, for which low pixel resoultion is the most prominent reason. Consequently, such pair of
values will be eliminated.
}
\references{
Kroener, S., Domenech Carbo, M.T., 2013. Determination of minimum pixel resolution for shape analysis: Proposal of a new data validation method for computerized images. Powder Technol. 245, 297 – 313.
}
\author{
Uwe Schnepf
}
