% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{clean_exclusion_size}
\alias{clean_exclusion_size}
\title{Function to exclude particles based on their size}
\usage{
clean_exclusion_size(data, lower.size, upper.size, diam)
}
\arguments{
\item{data}{A data frame that will be manipulated.}

\item{lower.size}{The lower size that is used as cut-off value.}

\item{upper.size}{The upper size that is used as cut-off value.}

\item{diam}{A vector inside the data that contains a measure of the particles diameter.}
}
\description{
All particles below a lower size limit and above an upper size limit are removed from the data set.
}
\author{
Uwe Schnepf
}
