% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MPA_utils.R
\name{compute_pathway_score}
\alias{compute_pathway_score}
\title{Computes a perturbation score - a metric that evaluates pathway disruption
based on the input metabolite data}
\usage{
compute_pathway_score(expr_data, pathway_data)
}
\arguments{
\item{expr_data}{the metabolite level data - data frame with metabolite average values for condition and control
rows are metabolites, rownames are metabolites ids and 2 columns condition and control}

\item{pathway_data}{the pathway data in a SBML format - this will be a custom structure}
}
\value{
numeric value the pathway perturbation score
}
\description{
Computes a perturbation score - a metric that evaluates pathway disruption
based on the input metabolite data
}
