% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MPA_utils.R
\name{compute_reaction_score}
\alias{compute_reaction_score}
\title{Computes a perturbation score - a metric that evaluates reaction disruption
based on the input metabolite data}
\usage{
compute_reaction_score(expr_data, reaction_data)
}
\arguments{
\item{expr_data}{the metabolite level data - data frame with metabolite average values for condition and control
rows are metabolites, rownames are metabolites ids and 2 columns condition and control}

\item{reaction_data}{the reaction data in a list format list
left and right with each list having another list of elements which
are the metabolites and its coefficient}
}
\value{
numeric value the reaction perturbation score
}
\description{
Computes a perturbation score - a metric that evaluates reaction disruption
based on the input metabolite data
}
