% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_function_details.R
\name{get_function_details}
\alias{get_function_details}
\title{Get Function Details}
\usage{
get_function_details(function_name)
}
\arguments{
\item{function_name}{Name of the test function (e.g., "F01", "F02", ..., "F23")}
}
\value{
A list containing:
\item{lb}{Lower bounds for the function}
\item{ub}{Upper bounds for the function}
\item{dim}{Number of dimensions}
\item{fobj}{The objective function}
}
\description{
Retrieves the details (bounds, dimension, and function object) for a specific
benchmark test function.
}
\examples{
# Get details for the Sphere function (F01)
details <- get_function_details("F01")
str(details)

}
