% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MPA_utils.R
\name{get_pathway_files}
\alias{get_pathway_files}
\title{Retrieves the files for the reactome files from url}
\usage{
get_pathway_files(reactome_url)
}
\arguments{
\item{reactome_url}{the reactome pathway data url}
}
\value{
named vector with the pathway file names
}
\description{
Retrieves the files for the reactome files from url
}
