% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/old.R
\name{graph_hist}
\alias{graph_hist}
\title{Histogram for particle size and shape distribution}
\usage{
graph_hist(
  name_file,
  polymer_type,
  this_descriptor,
  bin_number,
  name_axis,
  tick_Xaxis,
  tick_Yaxis = seq(0, 1000, by = 1000),
  limit_Xaxis = c(0, 1),
  limit_Yaxis = c(0, 1000)
)
}
\arguments{
\item{name_file}{A string to name the output images.}

\item{polymer_type}{A vector to select a certain kind of plastic.}

\item{this_descriptor}{A vector that contains the particle characteristic (e.g. size) to plot.}

\item{bin_number}{Sets the number of bins of the histogram.}

\item{name_axis}{A string to name the x-axis.}

\item{tick_Xaxis}{A pair of values to set the ticks of the x-axis.}

\item{tick_Yaxis}{A pair of values to set the ticks of the y-axis}

\item{limit_Xaxis}{A pair of values to specifiy the limit of the x-axis.}

\item{limit_Yaxis}{A pair of values to specifiy the limit of the y-axis.}
}
\description{
Plots a histogram with frequency and normalised frequency on the y-axis.
}
\references{
Filella, M., 2015. Questions of size and numbers in environmental research on microplastics: methodological and conceptual aspects. Environ. Chem. 12, 527.
}
\author{
Uwe Schnepf
}
