% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/old.R
\name{graph_KroenerCarbo}
\alias{graph_KroenerCarbo}
\title{Diagnostic plots for the Kroener & Carbo model}
\usage{
graph_KroenerCarbo(elon, circ, sphe, main = "", SE = FALSE, save = TRUE)
}
\arguments{
\item{elon}{A vector containing elongation values.}

\item{circ}{A vector containing circularity values.}

\item{sphe}{A vector containing sphericity values.}

\item{main}{A string to set a main title for the plot.}

\item{SE}{LOGICAL. Indicates whether a Sphericity-Elongation graph should be plotted or not. Default is set to FALSE.}

\item{save}{LOGICAL. Indicates whether plots should be automatically saved in an output folder or not. Default is set to TRUE.}
}
\description{
Implements an algorithm proposed by Kroener and Carbo to detect and delete falsely calculated shape factors
caused by digitization issues. It is based on a theoretical relationship between circularity/sphericity and elongtaion.
This relationship acts like a threshold for meausred pair of shape factor values. If they exceed the threshold, then something went
wrong in image processing, for which low pixel resoultion is the most prominent reason. Consequently, such pair of
values will be eliminated.
}
\references{
Kroener, S., Domenech Carbo, M.T., 2013. Determination of minimum pixel resolution for shape analysis: Proposal of a new data validation method for computerized images. Powder Technol. 245, 297 – 313.
}
\author{
Uwe Schnepf
}
