% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MPA_utils.R
\name{hello}
\alias{hello}
\title{Displays hello, world and the parameter x
based on the input metabolite data}
\usage{
hello(x)
}
\arguments{
\item{x}{the text to display next to Hello, world!}
}
\value{
string x the text that was displaied next to Hello, world!
}
\description{
Displays hello, world and the parameter x
based on the input metabolite data
}
