% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MPA_utils.R
\name{hello2}
\alias{hello2}
\title{Displays Hello again, world!
based on the input metabolite data}
\usage{
hello2()
}
\value{
string Hello again, world!
}
\description{
Displays Hello again, world!
based on the input metabolite data
}
