% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MPA_utils.R
\name{MPA_analysis}
\alias{MPA_analysis}
\title{Performs metabolic pathway analysis
based on the input metabolite data and the input pathways}
\usage{
MPA_analysis(expr_data_file, reactome_url)
}
\arguments{
\item{expr_data_file}{the file with the metabolite level data - tab separated file with
metabolite average values for condition and control
rows are metabolites, rownames are metabolites ids and 2 columns condition and control}

\item{reactome_url}{the reactome pathway data url}
}
\value{
data.frame with the ranked pathway and pathway scores and p-values
}
\description{
Performs metabolic pathway analysis
based on the input metabolite data and the input pathways
}
