% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MPA_utils.R
\name{read_expression_data}
\alias{read_expression_data}
\title{Retrieves from a tab-separated file a data.frame with the metabolite data}
\usage{
read_expression_data(file_name)
}
\arguments{
\item{file_name}{the metabolite level data file name - tab-separated file with
metabolite average values for condition and control
rows are metabolites, rownames are metabolites ids and 2 columns condition and control}
}
\value{
data.frame with the expression data
}
\description{
Retrieves from a tab-separated file a data.frame with the metabolite data
}
