% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MPA_utils.R
\name{read_pathway_data}
\alias{read_pathway_data}
\title{Reads pathway data from file}
\usage{
read_pathway_data(file_name)
}
\arguments{
\item{file_name}{of the metabolic pathway in SBML format}
}
\value{
list of reactions
}
\description{
Reads pathway data from file
}
