 \name{revista}
 \alias{revista}
 \title{Keywords of papers}
 \description{Keywords of papers
  }
 \usage{data(revista)}
 \format{Object whit text data was reading with leer.mpa("revista.txt",encoding="latin1")
}
 \source{Data from Revista Colombiana de Estadistica}
 \references{D. Rodriguez. Trabajo de Grado}  
  \keyword{datasets}

